requirements.txt
llama-cpp-python
huggingface_hub[cli]
gradio
duckdb